Extract file storage handlers from ctrl into service
