+++ apps/backend/package.json
@@ -116,6 +116,7 @@
     "untildify": "^5.0.0",
     "url-join": "^5.0.0",
     "url-parse": "^1.5.10",
+    "uuid": "^8.3.2",
     "yn": "^5.0.0",
     "yup": "^0.32.11"
   },
